requirements.txt
# PA-HCL: Physiology-Aware Hierarchical Contrastive Learning
# Environment: Python 3.10, CUDA 11.8, PyTorch 2.1.2

# ============== Core ==============
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2

# ============== Mamba (SSM) ==============
mamba-ssm>=1.2.0
causal-conv1d>=1.1.0

# ============== Distributed Training ==============
accelerate>=0.25.0

# ============== Audio Processing ==============
librosa>=0.10.1
soundfile>=0.12.1
scipy>=1.11.0
resampy>=0.4.2

# ============== Data & Config ==============
numpy>=1.24.0
pandas>=2.0.0
PyYAML>=6.0
omegaconf>=2.3.0
h5py>=3.9.0

# ============== Visualization ==============
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.4
scikit-learn>=1.3.0

# ============== Logging & Experiment Tracking ==============
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.66.0
rich>=13.5.0

# ============== Development ==============
pytest>=7.4.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0

# ============== Utilities ==============
einops>=0.7.0
timm>=0.9.7
